# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  product.CreateOptionValue:
    properties:
      name:
        type: string
    type: object
  product.CreateOptions:
    properties:
      name:
        type: string
      values:
        items:
          $ref: '#/definitions/product.CreateOptionValue'
        type: array
    type: object
  product.CreateProductRequest:
    properties:
      category_id:
        type: integer
      collection_ids:
        items:
          type: integer
        type: array
      files:
        items:
          $ref: '#/definitions/product.ProductFiles'
        type: array
      meta_description:
        type: string
      meta_title:
        type: string
      name:
        type: string
      options:
        items:
          $ref: '#/definitions/product.CreateOptions'
        type: array
      origin_price:
        minimum: 0
        type: integer
      sale_price:
        minimum: 0
        type: integer
      slug:
        type: string
      tags:
        items:
          type: string
        type: array
      variants:
        items:
          $ref: '#/definitions/product.CreateVariant'
        type: array
    required:
    - name
    - slug
    type: object
  product.CreateVariant:
    properties:
      file:
        type: string
      options:
        items:
          $ref: '#/definitions/product.VariantOption'
        type: array
      origin_price:
        type: integer
      sale_price:
        type: integer
      sku:
        type: string
      stock:
        type: integer
    type: object
  product.DeleteProductsRequest:
    properties:
      ids:
        items:
          type: integer
        type: array
    type: object
  product.OneProductResponse:
    properties:
      category_id:
        type: integer
      collections: {}
      files: {}
      id:
        type: integer
      is_active:
        type: boolean
      meta_description:
        type: string
      meta_title:
        type: string
      name:
        type: string
      options:
        items:
          $ref: '#/definitions/product.Option'
        type: array
      origin_price:
        type: integer
      sale_price:
        type: integer
      slug:
        type: string
      tags: {}
      variants:
        items:
          $ref: '#/definitions/product.OneVariant'
        type: array
    type: object
  product.OneVariant:
    properties:
      file:
        type: string
      id:
        type: integer
      options:
        items:
          $ref: '#/definitions/product.VariantOption'
        type: array
      origin_price:
        type: integer
      sale_price:
        type: integer
      sku:
        type: string
      stock:
        type: integer
    type: object
  product.Option:
    properties:
      id:
        type: integer
      name:
        type: string
      "no":
        type: integer
      product_id:
        type: integer
      values:
        items:
          $ref: '#/definitions/product.OptionValue'
        type: array
    type: object
  product.OptionValue:
    properties:
      id:
        type: integer
      name:
        type: string
      "no":
        type: integer
      option_id:
        type: integer
    type: object
  product.PaginatedResponse-any:
    properties:
      data:
        items: {}
        type: array
      page:
        example: 1
        type: integer
      page_size:
        example: 10
        type: integer
      total_items:
        example: 125
        type: integer
      total_pages:
        example: 13
        type: integer
    type: object
  product.ProductFiles:
    properties:
      is_primary:
        type: boolean
      name:
        type: string
      "no":
        type: integer
    type: object
  product.UpdateOptionValue:
    properties:
      id:
        type: integer
      name:
        type: string
    type: object
  product.UpdateOptions:
    properties:
      id:
        type: integer
      name:
        type: string
      values:
        items:
          $ref: '#/definitions/product.UpdateOptionValue'
        type: array
    type: object
  product.UpdateProductRequest:
    properties:
      category_id:
        type: integer
      collection_ids:
        items:
          type: integer
        type: array
      files:
        items:
          $ref: '#/definitions/product.ProductFiles'
        type: array
      meta_description:
        type: string
      meta_title:
        type: string
      name:
        type: string
      options:
        items:
          $ref: '#/definitions/product.UpdateOptions'
        type: array
      origin_price:
        minimum: 0
        type: integer
      sale_price:
        minimum: 0
        type: integer
      slug:
        type: string
      tags:
        items:
          type: string
        type: array
      variants:
        items:
          $ref: '#/definitions/product.UpdateVariants'
        type: array
    required:
    - name
    - slug
    type: object
  product.UpdateVariants:
    properties:
      file:
        type: string
      id:
        type: integer
      options:
        items:
          $ref: '#/definitions/product.VariantOption'
        type: array
      origin_price:
        type: integer
      sale_price:
        type: integer
      sku:
        type: string
      stock:
        type: integer
    type: object
  product.VariantOption:
    properties:
      option_id:
        type: integer
      option_name:
        type: string
      value:
        type: string
      value_id:
        type: integer
    type: object
...
info:
  contact: {}
  description: This is a sample server celler server.
  title: Swagger Example API
  version: "1.0"
paths:
...
  /products:
    delete:
      consumes:
      - application/json
      description: Deletes multiple products by their IDs
      parameters:
      - description: List of product IDs
        in: body
        name: ids
        required: true
        schema:
          $ref: '#/definitions/product.DeleteProductsRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Delete multiple products
      tags:
      - products
    get:
      description: Returns a list of products
      parameters:
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 10
        description: Page size
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/product.PaginatedResponse-any'
      security:
      - BearerAuth: []
      summary: Get product list
      tags:
      - products
    post:
      consumes:
      - application/json
      description: Creates a new product and returns the created product
      parameters:
      - description: Create product data
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/product.CreateProductRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Create a new product
      tags:
      - products
  /products/{id}:
    get:
      description: Returns a product by ID
      parameters:
      - description: Product ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/product.OneProductResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get a product
      tags:
      - products
    put:
      consumes:
      - application/json
      description: Updates a product
      parameters:
      - description: Product ID
        in: path
        name: id
        required: true
        type: integer
      - description: Update product data
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/product.UpdateProductRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Update a product
      tags:
      - products
...
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"